+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Change `Row::Base` to an abstract class
 
+### Misc
+
+* Change to use `Hash#each_pair` instead of `Hash#each`.
+
 
 ## 0.1.1 (2017-12-03)